### 2020-05-23 #2246 @whedon
[PRE REVIEW]: TTLocVis: A Twitter Topic Location Visualization Package
**Submitting author:** @xillig (<a href="http://orcid.org/0000-0003-2346-2841">Gillian Kant</a>)
**Repository:** <a href="https://github.com/xillig/TTLocVis" target ="_blank">https://github.com/xillig/TTLocVis</a>
**Version:** v0.1.0
**Reviewers:** @sara-02 , @linuxscout, @aozorahime
