### 2020-07-16 #2246 @trallard
Hi @ahurriyetoglu on behalf of the JOSS editorial team I would like to invite you to review this submission.

Would you be able and willing to take it?
