### 2020-07-23 #2507 @whedon
[REVIEW]: TTLocVis: A Twitter Topic Location Visualization Package
**Submitting author:** @xillig (<a href="http://orcid.org/0000-0003-2346-2841">Gillian Kant</a>)
**Repository:** <a href="https://github.com/xillig/TTLocVis" target ="_blank">https://github.com/xillig/TTLocVis</a>
**Version:** 1.0.4
**Reviewer:** @sara-02 , @linuxscout, @aozorahime
