### 2020-08-02 #2507 @sara-02
Hey @xillig
You may want to check out these projects/papers that have previousy worked on twiiter data vizulization and include them under the literature review for your paper.
Also, you need to point our how your package compares to existing work, maybe this list can be a starting point.
- https://www.csc2.ncsu.edu/faculty/healey/tweet_viz/
- Q.  Kong,  R.  Ram,  and  M.-A.  Rizoiu,  “A  toolkit  for  analyzing  andvisualizing  online  users  via  reshare  cascade  modeling,”ArXiv,  vol.abs/2006.06167, 2020.
- S.  Malik,  A.  Smith,  T.  Hawes,  P.  Papadatos,  J.  Li,  C.  Dunne,  andB. Shneiderman, “Topicflow: Visualizing topic alignment of twitter dataover time,” inASONAM, 2013, p. 720–726.
-  Florence   Ying   Wang,   A.   Sallaberry,  K.   Klein,   M.   Takatsuka,   andM.  Roche,  “Senticompass:  Interactive  visualization  for  exploring  andcomparing  the  sentiments  of  time-varying  twitter  data,”  inPacificVis,2015, pp. 129–133.
- M.  Hu,  K.  Wongsuphasawat,  and  J.  Stasko,  “Visualizing  social  mediacontent with sententree,”IEEE Transactions on Visualization and Com-puter Graphics, vol. 23, no. 1, pp. 621–630, 2017
