### 2020-08-03 #2507 @xillig
hi @sara-02
Thanks a lot for your suggestions. We are doing the review process with JOSS for the first time. Should we just incorporate your comments and then push a new version of the paper? Or how does this usually work? Thank you!
