### 2020-08-03 #2507 @xillig
Hi @trallard
Thanks for initiating the review process, I noticed that the issue contains three review check-lists, but @linuxscout is not an assignee of the issue. Will he also review the paper?
