### 2020-08-04 #2507 @sara-02
> hi @sara-02
> Thanks a lot for your suggestions. We are doing the review process with JOSS for the first time. Should we just incorporate your comments and then push a new version of the paper? Or how does this usually work? Thank you!

Hye @xillig there is no fixed process, you can keep incorporating the changes as they are suggested or wait for other reviewers to chime in as well. However, if we incorporate the changes on the go, then it will be easier for others when they start reviewing it, they will start from the improved version.
