### 2020-08-04 #2507 @aozorahime
Hi @xillig I also agree with @sara-02 about referring some works regarding twitter data visualization and there's no mention about installation process and describe how this software compares to other commonly-used packages. You may read or look for some JOSS papers which has been published so you can revise the paper directly.
