### 2020-08-06 #2507 @xillig
Hi @aozorahime
Thanks for the feedback. We will include the suggestions asap.
