+++ paper.md
@@ -42,7 +42,10 @@ space and time. The package can be used for a broad range of applications for sc
 topics discussed on Twitter. For instance, the package could be used to analyse the public discourse on the COVID-19 
 pandemic on Twitter in different countries and regions in the world over time. In particular, 
 data from the recently provided COVID-19 stream by Twitter can be analysed to research the discussion about COVID-19 
-on Twitter.^[https://developer.twitter.com/en/docs/labs/covid19-stream/overview]
+on Twitter.^[https://developer.twitter.com/en/docs/labs/covid19-stream/overview] In the following an overview of TTLocVis 
+will be provided. Finally, it will be also discussed how TTLocVis extends existing related software solutions. 
+The installation of TTLocVis can easily be done via pip. Further details on the installation and the package an be 
+found in the packages repository or on the documentation website of TTLocVis.^[https://ttlocvis.readthedocs.io/en/latest/#installation]
 
 In general, Topic Models are generative probabilistic models, that provide an insight into hidden information 
 in large text corpora by estimating the underlying topics of the texts in an unsupervised manner. In Topic Models, 
@@ -71,6 +74,93 @@ Above this, the spatial distribution of Tweets can be plotted on a world map, wh
 part of the world, in order to visualise the chosen sample of Tweets. As part of the mapping process, each Tweet is 
 classified by its most prevalent topic and colour coded (for example see figure Word Map 1 and figure World Map 2 for 
 the spatial distribution of the same selected topics at different points in time).
+
+A package that is most related to TTLocVis,  
+
+[Stojanovski2014] use their TweetViz framework is a similar fashion as we do: 
+They provide word clouds and topic distributions of twitter topics for a fed data set. 
+The main difference between TTLocVis and TweetViz that TTLocVis boosts its topical coherence 
+by tweet pooling. It also lets the user define the LDA topic number as well as automatically chooses
+ the best models in contrast to TweetViz. 
+
+-- Do they take the geolocation into account? No.
+-  How about the time dimension? Yes, but only for user hashtag usage.
+
+
+
+
+Twitter data can be analysed with a web application by [@malik2013] with a LDA Topic Model. The authors use
+so-called bins resembling time intervals of equal length. For each of these bins,
+a LDA model is estimated to account for the topical change over time. They then use cosine similarity to align the 
+topics from the several bins to a resulting topic. In contrast to this approach, the LDA model is trained on pooled 
+tweets in TTLocVis in order to improve the estimation results. The estimation procedure in [@malik2013] could 
+be beneficial in the modelling of topic changes in short time intervals. However, in this framework Topic Models are 
+estimated on very samples on which LDA Models usually do not perform well.  
+
+
+
+
+
+
+In [@hu20116] social media text is visualized with word clouds. In contrast to our approach, 
+the selected words are choosen by a mix of frequency and sentence structure rather than LDA Topic Models.  
+The package does not provide options for a visualisation of the spatial or time dimension. 
+
+ 
+
+
+Onorati et al., 2019 visualize Tweets Topics with their software focusing on word clouds, 
+tree maps and map visualization. Nevertheless, in contrast to TTLocVis, they do not estimate 
+their topics by LDA but rather use semantic relations and do focus on the contents of individual 
+tweets with regard to disaster-related classification.
+
+
+-- Do they take the geolocation into account? Yes, map visualization as mentioned (they plot the location of tweets 
+topicwise based on semantic relation). 
[…]
