+++ paper.bib
@@ -36,3 +36,55 @@
 	Title = {Latent Dirichlet Allocation},
 	Volume = {3},
 	Year = {2003}}
+
+
+
+
+@article{hu2016,
+  title={Visualizing social media content with sententree},
+  author={Hu, Mengdie and Wongsuphasawat, Krist and Stasko, John},
+  journal={IEEE transactions on visualization and computer graphics},
+  volume={23},
+  number={1},
+  pages={621--630},
+  year={2016},
+  publisher={IEEE}
+}
+
+@inproceedings{malik2013,
+  title={TopicFlow: visualizing topic alignment of Twitter data over time},
+  author={Malik, Sana and Smith, Alison and Hawes, Timothy and Papadatos, Panagis and Li, Jianyu and Dunne, Cody and Shneiderman, Ben},
+  booktitle={Proceedings of the 2013 IEEE/ACM international conference on advances in social networks analysis and mining},
+  pages={720--726},
+  year={2013}
+}
+
+@article{stojanovski2014,
+  title={Tweetviz: Twitter data visualization},
+  author={Stojanovski, Dario and Dimitrovski, Ivica and Madjarov, Gjorgji},
+  journal={Proceedings of the Data Mining and Data Warehouses},
+  volume={1},
+  number={2},
+  year={2014}
+}
+
+@article{onorati2019,
+  title={From social networks to emergency operation centers: A semantic visualization approach},
+  author={Onorati, Teresa and D{\'\i}az, Paloma and Carrion, Belen},
+  journal={Future Generation Computer Systems},
+  volume={95},
+  pages={829--840},
+  year={2019},
+  publisher={Elsevier}
+}
+
+@incollection{maia2016,
+  title={TweeProfiles3: visualization of spatio-temporal patterns on Twitter},
+  author={Maia, Andr{\'e} and Cunha, Tiago and Soares, Carlos and Abreu, Pedro Henriques},
+  booktitle={New Advances in Information Systems and Technologies},
+  pages={869--878},
+  year={2016},
+  publisher={Springer}
+}
+
+
