+++ paper.bib
@@ -88,3 +88,11 @@
 }
 
 
+@misc{kong2020,
+    title={A Toolkit for Analyzing and Visualizing Online Users via Reshare Cascade Modeling},
+    author={Quyu Kong and Rohit Ram and Marian-Andrei Rizoiu},
+    year={2020},
+    eprint={2006.06167},
+    archivePrefix={arXiv},
+    primaryClass={cs.SI}
+}
