+++ paper.md
@@ -79,8 +79,8 @@ To the knowledge of the authors no Python Package with a comparable functionalit
 A web tool that is most related to TTLocVis is TweetViz by [@Stojanovski2014]. TweetViz provides word clouds and topic 
 distributions for Twitter data. However, TTLocVis improves on TweetViz by optimizing the LDA input with tweet pooling and
 options for geo-spatial and temporal analysis. Further, a major limitation of TweetViz is that the number of topic for 
-the LDA estimation is always fixed as 20. TTLocVis gives the option to define a range of potential topic numbers and 
-also includes an algorithm to select the optional topic number according to coherence scoring. 
+the LDA estimation is always fixed at 20. TTLocVis gives the option to define a range of potential topic numbers and 
+also includes an algorithm to select the optimal topic number according to coherence scoring. 
 
 Alternatively, Twitter data may be analysed with a web application by [@malik2013] with a LDA Topic Model. 
 The authors use so-called bins resembling time intervals for the Topic Model estimation. For each of these bins,
@@ -95,7 +95,7 @@ word clouds, tree maps and map visualization. In contrast to TTLocVis, they do n
 their topics by LDA, but rather use semantic relations. The focus of their application is on the contents of individual 
 tweets with regard to disaster-related classification.
 
-In [@hu20116] a framework for social media text is provided. In contrast to our approach, 
+In [@hu2016] a framework for social media text is provided. In contrast to our approach, 
 the selected words are choosen by a mix of frequency and sentence structure rather than LDA Topic Models.  
 The package does not provide options for a visualisation of the spatial or time dimension. 
 
