+++ paper.md
@@ -95,12 +95,12 @@ word clouds, tree maps and map visualization. In contrast to TTLocVis, they do n
 their topics by LDA, but rather use semantic relations. The focus of their application is on the contents of individual 
 tweets with regard to disaster-related classification.
 
-In [@hu2016] a framework for social media text is provided. In contrast to our approach, 
-the selected words are choosen by a mix of frequency and sentence structure rather than LDA Topic Models.  
+In [@hu2016] a framework for social media text analysis is provided. In contrast to our approach, 
+the selected words are choosen by a mix of frequency and sentence structure rather than by LDA Topic Models.  
 The package does not provide options for a visualisation of the spatial or time dimension. 
 
-A recent analysis of COVID-19 related Tweets can be found in [@kong2020]. They draw on the package birdspotter and
-evently in order to analyse retweet cascades. Birdspotter is a package to analyse social influence and botness of 
+A recent analysis of COVID-19 related Tweets can be found in [@kong2020]. They use the packages Birdspotter and
+Evently in order to analyse retweet cascades. Birdspotter is a package to analyze social influence and Botness of 
 Twitter users, while Evently can be used to model the temporal spread of information. 
 
 
