+++ paper.bib
@@ -38,19 +38,6 @@
 	Year = {2003}}
 
 
-
-
-@article{hu2016,
-  title={Visualizing social media content with sententree},
-  author={Hu, Mengdie and Wongsuphasawat, Krist and Stasko, John},
-  journal={IEEE transactions on visualization and computer graphics},
-  volume={23},
-  number={1},
-  pages={621--630},
-  year={2016},
-  publisher={IEEE}
-}
-
 @inproceedings{malik2013,
   title={TopicFlow: visualizing topic alignment of Twitter data over time},
   author={Malik, Sana and Smith, Alison and Hawes, Timothy and Papadatos, Panagis and Li, Jianyu and Dunne, Cody and Shneiderman, Ben},
