+++ paper.bib
@@ -2,7 +2,7 @@
 %% http://bibdesk.sourceforge.net/
 
 
-%% Created for Christoph at 2020-05-22 21:24:27 +0200 
+%% Created for Christoph at 2020-08-12 17:11:07 +0200 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -10,18 +10,11 @@
 
 
 @inproceedings{Mehrotra,
-	Address = {New York, NY, USA},
 	Author = {Mehrotra, Rishabh and Sanner, Scott and Buntine, Wray and Xie, Lexing},
 	Booktitle = {Proceedings of the 36th International ACM SIGIR Conference on Research and Development in Information Retrieval},
 	Date-Added = {2020-05-22 19:28:07 +0200},
-	Date-Modified = {2020-05-22 21:24:27 +0200},
-	Isbn = {9781450320344},
-	Keywords = {microblogs, lda, topic modeling},
-	Location = {Dublin, Ireland},
-	Numpages = {4},
+	Date-Modified = {2020-08-12 17:11:06 +0200},
 	Pages = {889--892},
-	Publisher = {Association for Computing Machinery},
-	Series = {SIGIR '13},
 	Title = {Improving LDA Topic Models for Microblogs via Tweet Pooling and Automatic Labeling},
 	Year = {2013},
 	Bdsk-Url-1 = {https://doi.org/10.1145/2484028.2484166}}
@@ -29,57 +22,47 @@
 @article{blei,
 	Author = {Blei, David M. and Ng, Andrew Y. and Jordan, Michael I.},
 	Date-Added = {2020-05-22 19:25:54 +0200},
-	Date-Modified = {2020-05-22 19:28:45 +0200},
-	Issn = {1532-4435},
-	Journal = {J. Mach. Learn. Res.},
+	Date-Modified = {2020-08-12 17:08:47 +0200},
+	Journal = {Journal of  Machine Learning Research},
 	Pages = {993--1022},
 	Title = {Latent Dirichlet Allocation},
 	Volume = {3},
 	Year = {2003}}
 
-
 @inproceedings{malik2013,
-  title={TopicFlow: visualizing topic alignment of Twitter data over time},
-  author={Malik, Sana and Smith, Alison and Hawes, Timothy and Papadatos, Panagis and Li, Jianyu and Dunne, Cody and Shneiderman, Ben},
-  booktitle={Proceedings of the 2013 IEEE/ACM international conference on advances in social networks analysis and mining},
-  pages={720--726},
-  year={2013}
-}
+	Author = {Malik, Sana and Smith, Alison and Hawes, Timothy and Papadatos, Panagis and Li, Jianyu and Dunne, Cody and Shneiderman, Ben},
+	Booktitle = {Proceedings of the 2013 IEEE/ACM international conference on advances in social networks analysis and mining},
+	Pages = {720--726},
+	Title = {TopicFlow: visualizing topic alignment of Twitter data over time},
+	Year = {2013}}
 
-@article{stojanovski2014,
-  title={Tweetviz: Twitter data visualization},
-  author={Stojanovski, Dario and Dimitrovski, Ivica and Madjarov, Gjorgji},
-  journal={Proceedings of the Data Mining and Data Warehouses},
-  volume={1},
-  number={2},
-  year={2014}
-}
+@inproceedings{stojanovski2014,
+	Author = {Stojanovski, Dario and Dimitrovski, Ivica and Madjarov, Gjorgji},
+	Booktitle = {Proceedings of the Data Mining and Data Warehouses},
+	Date-Modified = {2020-08-12 17:10:23 +0200},
+	Pages = {1--4},
+	Title = {Tweetviz: Twitter data visualization},
+	Year = {2014}}
 
 @article{onorati2019,
[…]
