### 2020-08-12 #2507 @xillig
Hi @sara-02 and @aozorahime,
we uploaded an updated version of our paper, embedding your suggestions.
