### 2020-09-14 #2507 @xillig
Dear @sara-02, @aozorahime, @linuxscout and @trallard,
we hope that you are all doing well. Can we provide any further assistance regarding the open points? Please also consult the documentation website (https://ttlocvis.readthedocs.io/en/latest/) if you may have questions or write us at any time.
