### 2020-10-06 #2507 @xillig
Dear @trallard, @sara-02, @aozorahime,
We hope that you are all well and not too severely affected by the current Covid situation. For us a decision about our paper is quiet important in terms of funding applications. Given that the extended review timeline of six weeks is already past since a while, we would kindly like to ask you if you could give us an indication as to when we can expect a conclusion of the review?

Thank you!
xillig
