### 2020-10-14 #2507 @sara-02
@xillig rest of the features and documentation are well put, but where are the commands to run the test listed?
I see you have a test folder but how to operate on it?
