### 2020-10-14 #2507 @sara-02
As we have to cover that under `Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?`
