### 2020-10-14 #2507 @sara-02
Also, I feel we can better structure the content of the paper instead of putting everythhing under the summary heading.
