### 2020-10-14 #2507 @sara-02
So, here is a suggested breakdown:
```
Summary: Para 1
About the package: para 2(In general...) to para 5 (Above this, the spatial distribution...)
Comparison with existing tools: Para 6 onwards
```
