### 2020-10-14 #2507 @xillig
Hi @sara-02,
you can find the guide for how to run the example files in the documentation under the following link: https://ttlocvis.readthedocs.io/en/latest/examples/
Please note that you can't use the TwitterStreamer without an Twitter Developer account, all other functions are usable with the example data. Please also note that the example data will not return any meaningful analytical results since it is for functional demonstration and tutorial purposes only.

Regarding your suggested breakdown, we will review the paper in regards to your feedback asap.

Thank you and all the best!
