### 2020-10-14 #2507 @xillig
Hey, @sara-02,
Thanks for the clarification,
>
>
> As we have to cover that under `Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?`

There is no automated unit testing available, since our background is statistics and not computer science, so that we are not firm with unit testing. That's why we decided for the second option to provide manual testing steps as outlined in the "Usage and Example" guideline, to explain how all functions are used.

>
>
> Hey, I have reviewed this documentation, my concern is regarding running unit tests.
> Is there a way we can test the modules with some dummy twitter data?

Using the example data provided (which can be found here: https://github.com/xillig/TTLocVis/tree/master/example_files), you can   run all the features of TTLocVis except for the TwitterStreamer class (Which was produced by the TwitterStreamer class). For example you can test the Cleaner class:

from TTLocVis.module import Cleaner
c = Cleaner(load_path=r'path-to-a-JSON-file-folder')
print(c.raw_data)
c.saving(r'path-to-where-the-cleaned-data-shall-be-saved')

where r'path-to-a-JSON-file-folder' should point to the example folder.

I hope this answers your questions
