### 2020-10-15 #2507 @xillig
Dear @trallard, @sara-02, @aozorahime,
 thank you very much for completing the review and for all the great comments. This really helped a lot to improve the paper and we really appreciate that you invested so much time despite of the COVID-19 situation.

 @trallard: What are the next steps in the publication process?

On the JOSS Website we find the following instructions:
>
>
> Upon successful completion of the review, authors will make a tagged release of the software, and deposit a copy of the repository with a data-archiving service such as Zenodo or figshare, get a DOI for the archive, and update the review issue thread with the version number and DOI.

 Can we proceed with this?

Thank you all very much!
