### 2020-10-19 #2507 @xillig
Hi @tallard,
now that @sara-02 and @aozorahime completed their reviews and helped us to really improve TTLocVis a lot, it would really great to make further progress with the review. How can we proceed? What are the next steps? Also @sara-02 and @aozorahime do you perhaps know what to do now?

Thank you all!

xillig
