### 2020-10-20 #2507 @xillig
Hi @trallard,
we already created a tagged release on Zenodo. This is the DOI: https://zenodo.org/badge/latestdoi/258153686

Additionally, we would like to point out once more, that @linuxscout might be assigned as a Reviewer (he/she has a checklist above) but never has been a participant in this issue and probably never reviewed the project. Nevertheless, we do not think this is a problem since there are only two reviews nessesary to move forward, which are by now successfully completed by @sara-02 and @aozorahime.

We hope this helps to move the case forward.
Thank you!
