### 2020-10-21 #2507 @xillig
Hi @linuxscout,
thanks for your message. We agree that no further (third) review is necessary and your valueable work as a reviewer is more important for the numerous other submissions to JOSS!

Thus if @trallard agrees we would proceed with the publication process.
