### 2020-10-21 #2507 @trallard
Hey all thanks for your patience and your time doing the reviews. I am going to move forward with the next steps.

@xillig can you please make sure that the software version is up to date and confirm it here?
