### 2020-10-21 e67f3a7
:pencil: Grammatical and typo edits to paper
