+++ paper.md
@@ -27,7 +27,7 @@ affiliations:
  - name: Centre for Statistics, Georg-August-Universität Göttingen, Germany
    index: 1
  - name: Campus-Institut Data Science, Göttingen, Germany
-   index: 2  
+   index: 2
 date: 25 May 2020
 bibliography: paper.bib
 
@@ -36,76 +36,54 @@ bibliography: paper.bib
 # Summary
 
 The package TTLocVis provides a broad range of methods to generate, clean, analyze and visualize the contents of Twitter
-data. TTLocVis enables the user to work with geo-spatial Twitter data and to generate topic distributions from Latent 
-Dirichlet Allocation (LDA) Topic Models [@blei] for geo-coded Tweets. As such, TTLocVis is an innovative 
-tool to work with geo-coded text on a high geo-spatial resolution to analyse the public discourse on various topics in 
-space and time. The package can be used for a broad range of applications for scientific research to gain insights into 
-topics discussed on Twitter. For instance, the package could be used to analyse the public discourse on the COVID-19 
-pandemic on Twitter in different countries and regions in the world over time. In particular, 
-data from the recently provided COVID-19 stream by Twitter can be analysed to research the discussion about COVID-19 
-on Twitter.^[https://developer.twitter.com/en/docs/labs/covid19-stream/overview] In the following an overview of TTLocVis 
-will be provided. Finally, it will be also discussed how TTLocVis extends existing related software solutions. 
-The installation of TTLocVis can easily be done via pip. Further details on the installation and the package an be 
+data. TTLocVis enables the user to work with geo-spatial Twitter data and to generate topic distributions from Latent
+Dirichlet Allocation (LDA) Topic Models [@blei] for geo-coded Tweets. As such, TTLocVis is an innovative tool to work with geo-coded text on a high geo-spatial resolution to analyze the public discourse on various topics in space and time. The package can be used for a broad range of applications for scientific research to gain insights into topics discussed on Twitter. For instance, the package could be used to analyze the public discourse on the COVID-19
+pandemic on Twitter in different countries and regions in the world over time. In particular,
+data from the recently provided COVID-19 stream by Twitter can be analyzed to research the discussion about COVID-19
+on Twitter.^[https://developer.twitter.com/en/docs/labs/covid19-stream/overview] In the following, an overview of TTLocVis will be provided. Finally, it will also be discussed how TTLocVis extends existing related software solutions.
+The installation of TTLocVis can easily be done via pip. Further details on the installation and the package can be
 found in the packages repository or on the documentation website of TTLocVis.^[https://ttlocvis.readthedocs.io/en/latest/#installation]
 
 ## About the package
 
-In general, Topic Models are generative probabilistic models, that provide an insight into hidden information 
-in large text corpora by estimating the underlying topics of the texts in an unsupervised manner. In Topic Models, 
-each topic is a distribution over words that can be labeled by humans. For the purpose of labelling histograms and 
-word clouds (for example see graph) provide helpful visualizations for the decision-making process of the user [@blei].
+In general, Topic Models are generative probabilistic models, that provide an insight into hidden information in large text corpora by estimating the underlying topics of the texts in an unsupervised manner. In Topic Models,
+each topic is a distribution over words that can be labeled by humans. For the purpose of labelling histograms and word clouds (for example, see graph) provide helpful visualizations for the decision-making process of the user [@blei].
 
 Firstly, the package allows the user to collect Tweets using a Twitter developer account for any area in the world.
-Subsequently, the inherently noisy Twitter data can be cleaned, transformed and exported. 
-In particular, TTLocVis enables the user to apply LDA Topic Models on extremely sparse Twitter data by preparing 
[…]
