### 2020-10-21 56fce3b
:pencil: Grammatical and typo edits to paper
