### 2020-10-21 #2507 @trallard
@xillig  there seem to be some DOIs missing, would you be able to check these, please?

Also, I sent a PR with some edits <https://github.com/xillig/TTLocVis/pull/3>
