+++ paper.bib
@@ -17,7 +17,7 @@
 	Pages = {889--892},
 	Title = {Improving LDA Topic Models for Microblogs via Tweet Pooling and Automatic Labeling},
 	Year = {2013},
-	Bdsk-Url-1 = {https://doi.org/10.1145/2484028.2484166}}
+    doi = {10.1145/2484028.2484166}}
 
 @article{blei,
 	Author = {Blei, David M. and Ng, Andrew Y. and Jordan, Michael I.},
@@ -50,7 +50,9 @@
 	Journal = {Future Generation Computer Systems},
 	Pages = {829--840},
 	Title = {From social networks to emergency operation centers: A semantic visualization approach},
-	Year = {2019}}
+	doi = {10.1016/j.future.2018.01.052}
+	Year = {2019}
+	}
 
 @incollection{maia2016,
 	Author = {Maia, Andr{\'e} and Cunha, Tiago and Soares, Carlos and Abreu, Pedro Henriques},
@@ -58,6 +60,7 @@
 	Pages = {869--878},
 	Publisher = {Springer},
 	Title = {TweeProfiles3: visualization of spatio-temporal patterns on Twitter},
+	doi = {10.1007/978-3-319-31232-3_82},
 	Year = {2016}}
 
 @misc{kong2020,
