+++ paper.bib
@@ -50,7 +50,7 @@
 	Journal = {Future Generation Computer Systems},
 	Pages = {829--840},
 	Title = {From social networks to emergency operation centers: A semantic visualization approach},
-	doi = {10.1016/j.future.2018.01.052}
+	doi = {10.1016/j.future.2018.01.052},
 	Year = {2019}
 	}
 
