+++ paper.bib
@@ -54,15 +54,6 @@
 	Year = {2019}
 	}
 
-@incollection{maia2016,
-	Author = {Maia, Andr{\'e} and Cunha, Tiago and Soares, Carlos and Abreu, Pedro Henriques},
-	Booktitle = {New Advances in Information Systems and Technologies},
-	Pages = {869--878},
-	Publisher = {Springer},
-	Title = {TweeProfiles3: visualization of spatio-temporal patterns on Twitter},
-	doi = {10.1007/978-3-319-31232-3_82},
-	Year = {2016}}
-
 @misc{kong2020,
 	Archiveprefix = {arXiv},
 	Author = {Quyu Kong and Rohit Ram and Marian-Andrei Rizoiu},
