### 2020-10-21 #2507 @xillig
Dear @trallard,
Thank you for all your efforts. We included the DOIs of the references and fixed authorship issue in zenodo by making a new release. We also merged your changes into master. We can also confirm that the software version is up to date: The version number is 1.0.4, with the same DOI as before: https://zenodo.org/badge/latestdoi/258153686

Thanks to everyone for their contributions.

All the best,
Gillian, Christoph and Benjamin
