### 2020-10-25 #2507 @kyleniemeyer
Hi @xillig, I'm the EIC on duty this week, doing some final checks before publishing. Can

- [ ] Please edit the Zenodo archive author list to match the paper—it looks like @trallard got included!
- [ ] Your article is missing an explicit [Statement of Need section](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), which we have begun requiring for articles. Can you please add? You may be able to use existing text in the paper.
- [ ] In the last paragraph of the "Comparison with existing tools" section, I think there is an extraneous comma before the citation.
- [ ] The Kong 2020 reference seems to be missing some details
- [ ] Are there DOIs available for the other references?
