+++ paper.bib
@@ -26,6 +26,7 @@
 	Journal = {Journal of  Machine Learning Research},
 	Pages = {993--1022},
 	Title = {Latent Dirichlet Allocation},
+	doi = {10.1162/jmlr.2003.3.4-5.993},
 	Volume = {3},
 	Year = {2003}}
 
@@ -34,6 +35,7 @@
 	Booktitle = {Proceedings of the 2013 IEEE/ACM international conference on advances in social networks analysis and mining},
 	Pages = {720--726},
 	Title = {TopicFlow: visualizing topic alignment of Twitter data over time},
+	doi={10.1145/2492517.2492639},
 	Year = {2013}}
 
 @inproceedings{stojanovski2014,
@@ -54,10 +56,11 @@
 	Year = {2019}
 	}
 
-
-@misc{kong2020,
-	Archiveprefix = {arXiv},
-	Author = {Quyu Kong and Rohit Ram and Marian-Andrei Rizoiu},
-	Date-Modified = {2020-08-12 17:10:48 +0200},
-	Title = {A Toolkit for Analyzing and Visualizing Online Users via Reshare Cascade Modeling},
-	Year = {2020}}
+@article{kong2020,
+      title={evently: A Toolkit for Analyzing Online Users via Reshare Cascade Modeling},
+      author={Quyu Kong and Rohit Ram and Marian-Andrei Rizoiu},
+      year={2020},
+      eprint={2006.06167},
+      archivePrefix={arXiv},
+      primaryClass={cs.SI}
+}
