### 2020-10-26 #2507 @xillig
Hi @kyleniemeyer,

thank you very much for your feedback. Unfortunatly, for the two references that have no DOIs now, there are none available to your knowledge. Appart from that, we have incorporated all of your suggested points. Please let us know if we can do anything else or move forward.

Thanks to everyone!
