### 2020-10-26 #2507 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
